hi-res png export for 3D view
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@39387 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/geogebra3D/euclidian3D/openGL/ManagerShaders.java
@@ -610,7 +610,7 @@ public class ManagerShaders extends Manager {
 	}
 
 	@Override
-	protected void vertexInt(int x, int y, int z) {
+	protected void vertexInt(double x, double y, double z) {
 		vertex(x, y, z);
 	}
 
@@ -649,7 +649,8 @@ public class ManagerShaders extends Manager {
 	}
 
 	@Override
-	public void rectangleGeometry(int x, int y, int z, int width, int height) {
+	public void rectangleGeometry(double x, double y, double z, double width,
+			double height) {
 
 		startGeometry(Manager.Type.TRIANGLE_STRIP);
 		texture(0, 0);
